TFS1165747 Separated samples and added readmes
