+++ samples/Java/Basic/SendBatch/README.md
@@ -1,0 +1,45 @@
+# Send batch events to Azure Event Hubs using Java
+
+The [Send events to Azure Event Hubs using Java](https://docs.microsoft.com/azure/event-hubs/event-hubs-java-get-started-send) tutorial walks you through ingesting into your event hub using Java with this code.
+
+This sample differs from the basic tutorial in that it collects a set of events and sends them all at once in a single gesture:
+
+``` Java
+    final LinkedList<EventData> events = new LinkedList<>();
+    for (int count = 1; count < 11; count++) {
+        final PayloadEvent payload = new PayloadEvent(count);
+        final byte[] payloadBytes = gson.toJson(payload).getBytes(Charset.defaultCharset());
+        final EventData sendEvent = new EventData(payloadBytes);
+        sendEvent.getProperties().put("from", "javaClient");
+        events.add(sendEvent);
+    }
+
+    sender.sendSync(events);
+```
+
+To run the sample, you need to edit the [sample code](src/main/java/com/microsoft/azure/eventhubs/samples/sendbatch/SendBatch.java) and provide the following information:
+
+```java
+    final String namespaceName = "----ServiceBusNamespaceName-----";
+    final String eventHubName = "----EventHubName-----";
+    final String sasKeyName = "-----SharedAccessSignatureKeyName-----";
+    final String sasKey = "---SharedAccessSignatureKey----";
+```
+
+## Prerequisites
+
+Please refer to the [overview README](../../readme.md) for prerequisites and setting up the sample environment, including creating an Event Hubs cloud namespace and an Event Hub. 
+
+## Build and run
+
+The sample can be built independently with 
+
+```bash
+mvn clean package
+```
+
+and then run with (or just from VS Code or another Java IDE)
+
+```bash
+java -jar ./target/sendbatch-1.0.0-jar-with-dependencies.jar
+```
