+++ samples/Java/pom.xml
@@ -1,36 +1,33 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 
-  	<modelVersion>4.0.0</modelVersion>
 
-	<artifactId>azure-eventhubs-samples</artifactId>
-	<name>azure-eventhubs-samples</name>
+    <modelVersion>4.0.0</modelVersion>
+    <groupId>com.microsoft.azure</groupId>
+    <artifactId>azure-eventhubs-samples</artifactId>
+    <name>azure-eventhubs-samples</name>
+    <version>1.0.0</version>
+    <packaging>pom</packaging>
 
-	<dependencies>
-	 	<dependency>
-	  		<groupId>com.microsoft.azure</groupId>
-	  		<artifactId>azure-eventhubs-eph</artifactId>
-	  		<version>0.15.0</version>
-	 	</dependency>
-	</dependencies>
+    <properties>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+    </properties>
+    <build>
+        <directory>${project.basedir}/target</directory>
+        <outputDirectory>${project.build.directory}/classes</outputDirectory>
+        <finalName>${project.artifactId}-${project.version}</finalName>
+        <testOutputDirectory>${project.build.directory}/classes</testOutputDirectory>
+        <sourceDirectory>${project.basedir}/src/</sourceDirectory>
+        <testSourceDirectory>${project.basedir}/test/</testSourceDirectory>
+    </build>
 
-	<build>
-		<plugins>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-compiler-plugin</artifactId>
-				<version>3.5</version>
-				<configuration>
-					<source>1.8</source>
-					<target>1.8</target>
-					<optimize>true</optimize>
-					<showDeprecation>true</showDeprecation>
-					<showWarnings>true</showWarnings>
-				</configuration>
-			</plugin>
-		</plugins>
-	</build>
-
-	<groupId>com.microsoft.azure</groupId>
-	<version>1.0.0</version>
-
+    <modules>
+        <module>Basic/EventProcessorSample</module>
+        <module>Basic/ReceiveByDateTime</module>
+        <module>Basic/ReceiveUsingOffset</module>
+        <module>Basic/Send</module>
+        <module>Basic/SendBatch</module>
+        <module>Benchmarks/AutoScaleOnIngress</module>
+        <module>Benchmarks/IngressBenchmark</module>
+    </modules>
 </project>
