--- samples/Java/send.md
@@ -1,3 +1,0 @@
-# Send events to Azure Event Hubs using Java
-
-**Note**: This sample is available as a tutorial [here](https://docs.microsoft.com/azure/event-hubs/event-hubs-java-get-started-send).
